AI Laywer/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.43
aiosqlite==0.21.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.10.1
email-validator==2.2.0
httpx==0.28.1
reportlab==4.4.3
groq==0.31.0
google-generativeai==0.8.5
pdfplumber==0.11.7
pypdf==5.9.0
pillow==11.3.0
pytesseract==0.3.13
supabase==2.18.1
boto3==1.40.29
razorpay==1.4.2
